mob -> long

+++ Assets/Scripts/Assets.src.g/BigBoss.cs
@@ -114,7 +114,7 @@ namespace Assets.src.g
 
 		public new int forceWait;
 
-		public BigBoss(int id, short px, short py, int templateID, double hp, double maxhp, int s)
+		public BigBoss(int id, short px, short py, int templateID, long hp, long maxhp, int s)
 		{
 			xFirst = (x = px + 20);
 			yFirst = (y = py);
@@ -249,38 +249,38 @@ namespace Assets.src.g
 			updateShadown();
 			switch (status)
 			{
-			case 2:
-				updateMobStandWait();
-				break;
-			case 4:
-				timeStatus = 0;
-				updateMobFly();
-				break;
-			case 3:
-				updateMobAttack();
-				break;
-			case 5:
-				timeStatus = 0;
-				updateMobWalk();
-				break;
-			case 6:
-				timeStatus = 0;
-				p1++;
-				y += p1;
-				if (y >= yFirst)
-				{
-					y = yFirst;
-					p1 = 0;
-					status = 5;
-				}
-				break;
-			case 7:
-				updateInjure();
-				break;
-			case 0:
-			case 1:
-				updateDead();
-				break;
+				case 2:
+					updateMobStandWait();
+					break;
+				case 4:
+					timeStatus = 0;
+					updateMobFly();
+					break;
+				case 3:
+					updateMobAttack();
+					break;
+				case 5:
+					timeStatus = 0;
+					updateMobWalk();
+					break;
+				case 6:
+					timeStatus = 0;
+					p1++;
+					y += p1;
+					if (y >= yFirst)
+					{
+						y = yFirst;
+						p1 = 0;
+						status = 5;
+					}
+					break;
+				case 7:
+					updateInjure();
+					break;
+				case 0:
+				case 1:
+					updateDead();
+					break;
 			}
 		}
